Now a single minimization can be done in C++ only with 2 Rigidbody and Attract forcefield.
* added L-BFGS minimizer interface.
* added driver{1-3}.f which are good examples for L-BGFS
* SConstruct could now generate a static library for C++

- Python interface not complete for the new classes

+++ rigidbody.cpp
@@ -39,7 +39,7 @@ Coord3D Rigidbody::FindCenter() const
     {
         center =  center + GetCoords(i);
     }
-    return ( (1.0/this->Size())*center);
+    return ( (1.0/(double)this->Size())*center);
 }
